+++ lib/src/event_controller.dart
@@ -6,9 +6,7 @@ import 'dart:collection';
 
 import 'package:flutter/material.dart';
 
-import 'calendar_event_data.dart';
-import 'extensions.dart';
-import 'typedefs.dart';
+import '../calendar_view.dart';
 
 class EventController<T extends Object?> extends ChangeNotifier {
   /// Calendar controller to control all the events related operations like,
@@ -74,6 +72,42 @@ class EventController<T extends Object?> extends ChangeNotifier {
   //#endregion
 
   //#region Public Methods
+  /// Deletes a recurring event based on the specified deletion type.
+  ///
+  /// This method handles the deletion of recurring events by determining the
+  /// type of deletion
+  /// requested (all events, the current event, or following events) and
+  /// performing the appropriate action.
+  ///
+  /// Takes the following parameters:
+  /// - [date]: The date of the event to be deleted.
+  /// - [event]: The event data to be deleted.
+  /// - [deleteEventType]: The `DeleteEventType` of deletion to perform
+  /// (all events, the current event, or following events).
+  ///
+  /// The method performs the following actions based on the [deleteEventType]:
+  /// - [DeleteEvent.all]: Removes the entire series of events.
+  /// - [DeleteEvent.current]: Deletes only the current event.
+  /// - [DeleteEvent.following]: Deletes the current event and
+  /// all subsequent events.
+  void deleteRecurrenceEvent({
+    required DateTime date,
+    required CalendarEventData<T> event,
+    required DeleteEvent deleteEventType,
+  }) {
+    switch (deleteEventType) {
+      case DeleteEvent.all:
+        remove(event);
+        break;
+      case DeleteEvent.current:
+        _deleteCurrentEvent(date, event);
+        break;
+      case DeleteEvent.following:
+        _deleteFollowingEvents(date, event);
+        break;
+    }
+  }
+
   /// Add all the events in the list
   /// If there is an event with same date then
   void addAll(List<CalendarEventData<T>> events) {
@@ -142,6 +176,10 @@ class EventController<T extends Object?> extends ChangeNotifier {
         includeFullDayEvents: includeFullDayEvents);
   }
 
+  /// Returns repeated events on given date.
+  List<CalendarEventData<T>> getRecurringEventsOnDay(DateTime date) =>
+      _calendarData.getRecurringEventsOnDay(date);
+
   /// Returns full day events on given day.
   List<CalendarEventData<T>> getFullDayEvent(DateTime date) {
     return _calendarData.getFullDayEvent(date.withoutTime);
@@ -157,6 +195,36 @@ class EventController<T extends Object?> extends ChangeNotifier {
     }
   }
   //#endregion
+
+//#region Private Methods
+  void _deleteCurrentEvent(DateTime date, CalendarEventData<T> event) {
+    final excludeDates = event.recurrenceSettings?.excludeDates ?? []
+      ..add(date);
+    final updatedRecurrenceSettings =
+        event.recurrenceSettings?.copyWith(excludeDates: excludeDates);
+    final updatedEvent =
+        event.copyWith(recurrenceSettings: updatedRecurrenceSettings);
+    update(event, updatedEvent);
+  }
+
+  /// If the selected date to delete the event is the same as the event's start date, delete all recurrences.
+  /// Otherwise, delete the event on the selected date and all subsequent recurrences.
+  void _deleteFollowingEvents(DateTime date, CalendarEventData<T> event) {
+    final newEndDate = date.subtract(
+      const Duration(days: 1),
+    );
+    final updatedRecurrenceSettings = event.recurrenceSettings?.copyWith(
+      endDate: newEndDate,
+    );
+    if (date == event.date) {
+      remove(event);
+    } else {
+      final updatedEvent =
+          event.copyWith(recurrenceSettings: updatedRecurrenceSettings);
+      update(event, updatedEvent);
+    }
+  }
+//#endregion
 }
 
 /// Stores the list of the calendar events.
@@ -217,6 +285,10 @@ class CalendarData<T extends Object?> {
   UnmodifiableListView<CalendarEventData<T>> get fullDayEventList =>
       UnmodifiableListView(_fullDayEventList);
 
+  /// Stores all recurring events
+  final _recurringEventsList = <CalendarEventData<T>>[];
+  //#endregion
+
   //#region Data Manipulation Methods
   void addFullDayEvent(CalendarEventData<T> event) {
     // TODO: add separate logic for adding full day event and ranging event.
@@ -239,7 +311,6 @@ class CalendarData<T extends Object?> {
     } else {
       _singleDayEvents[date]!.addEventInSortedManner(event, _eventSorter);
     }
-
     _eventList.add(event);
   }
 
@@ -249,6 +320,12 @@ class CalendarData<T extends Object?> {
 
     // TODO: improve this...
     if (_eventList.contains(event)) return;
+
+    if (event.isRecurringEvent) {
+      _eventList.add(event);
+      _recurringEventsList.add(event);
+      return;
+    }
 
     if (event.isFullDayEvent) {
       addFullDayEvent(event);
@@ -278,6 +355,12 @@ class CalendarData<T extends Object?> {
   }
 
   void removeEvent(CalendarEventData<T> event) {
+    if (event.isRecurringEvent) {
+      _eventList.remove(event);
+      _recurringEventsList.remove(event);
+      return;
+    }
+
     if (event.isFullDayEvent) {
       removeFullDayEvent(event);
     } else if (event.isRangingEvent) {
@@ -311,6 +394,151 @@ class CalendarData<T extends Object?> {
   }
   //#endregion
 
+  //#region Helper Methods
+  // Event is not recurring. So, no need to handle recurrence.
+  // This method checks for the event whether it should exclude
+  // or not on given date.
+  // Event is excluded:
+  // - If given date is of before event start date
+  // - If given date is after recurrence end date
+  // - If given date is in excluded list of events.
+  // On returning true it excludes event and on false it won't exclude.
+  bool _isExcluded(RecurrenceSettings settings, DateTime date) {
+    final recurrenceEndDate = settings.endDate;
+    return (recurrenceEndDate != null && date.isAfter(recurrenceEndDate)) ||
+        (settings.excludeDates?.contains(date) ?? false);
+  }
+
+  /// Determines whether the given date should be included as a recurring event
+  /// for daily recurrence settings.
+  ///
+  /// Returns `true` if the event should repeat on the given `currentDate`,
+  /// otherwise returns `false`.
+  /// `endDate` may change, such as when handling deletions or updates.
+  ///
+  /// - If `recurrenceEndDate` is not specified, the event repeats indefinitely
+  ///   and this method returns `true`.
+  /// - If `recurrenceEndDate` is specified:
+  ///   - The event is included if the `currentDate` is before
+  ///   the `recurrenceEndDate`.
+  ///   - The event is also included on the exact `recurrenceEndDate`
+  ///     (checked using `isAtSameMomentAs`), allowing the event to occur
+  ///     on the last day.
+  bool _isDailyRecurrence({
+    required DateTime currentDate,
+    required RecurrenceSettings recurrenceSettings,
+  }) {
+    final recurrenceEndDate = recurrenceSettings.endDate;
+    return recurrenceEndDate == null ||
+        (currentDate.isBefore(recurrenceEndDate) ||
+            currentDate.isAtSameMomentAs(recurrenceEndDate));
+  }
+
+  /// If the weekday matches with `recurrenceSettings` and there is no end date,
+  /// the recurrence is infinite
+  ///
+  ///
+  /// If the weekday matches and there is an end date, check if the current date
+  /// is before or on the end date
+  /// This ensures the recurrence continues until the specified end date
+  ///
+  /// Recurrence endDate may change if event is deleted.
+  bool _isWeeklyRecurrence({
+    required DateTime currentDate,
+    required RecurrenceSettings recurrenceSettings,
+  }) {
+    // Adjust weekday to zero-based indexing and
+    // check if date’s weekday is in the recurrence weekdays
+    final isMatchingWeekday =
+        recurrenceSettings.weekdays.contains(currentDate.weekday - 1);
+    final recurrenceEndDate = recurrenceSettings.endDate;
+
+    if (!isMatchingWeekday) {
+      return false;
+    }
+
+    // If no end date is specified, repeat infinitely
+    return recurrenceEndDate == null ||
+        (currentDate.isBefore(recurrenceEndDate) ||
+            currentDate.isAtSameMomentAs(recurrenceEndDate));
+  }
+
+  // Repeat event on same day
+  // Returns true if event should repeat on the given date otherwise false.
+  // For monthly repetition of event event start date & given date should match.
+  // repetition will include the recurrence end date.
+  bool _isMonthlyRecurrence({
+    required DateTime currentDate,
+    required DateTime startDate,
+    required RecurrenceSettings recurrenceSettings,
+  }) {
+    // Exclude if day is different
+    if (currentDate.day != startDate.day) {
+      return false;
+    }
+
+    // Continues if day is same
+    final recurrenceEndDate = recurrenceSettings.endDate;
+
+    return recurrenceEndDate == null ||
+        (currentDate.isBefore(recurrenceEndDate) ||
+            currentDate.isAtSameMomentAs(recurrenceEndDate));
+  }
+
+  // If end date is not mentioned repeat infinitely
+  // If end date is mentioned repeat till end date including last date
+  // End date will change in case of "Following events" are deleted
+  bool _isYearlyRecurrence({
+    required DateTime currentDate,
+    required DateTime startDate,
+    required RecurrenceSettings recurrenceSettings,
+  }) {
+    if (currentDate.month != startDate.month ||
+        currentDate.day != startDate.day) {
+      return false;
+    }
+
+    final recurrenceEndDate = recurrenceSettings.endDate;
+    return recurrenceEndDate == null ||
+        (currentDate.isBefore(recurrenceEndDate) ||
+            currentDate.isAtSameMomentAs(recurrenceEndDate));
+  }
+
+  bool _handleRecurrence({
+    required DateTime currentDate,
+    required DateTime eventStartDate,
+    required DateTime eventEndDate,
+    required RecurrenceSettings recurrenceSettings,
+  }) {
+    switch (recurrenceSettings.frequency) {
+      case RepeatFrequency.doNotRepeat:
+        return currentDate.isAtSameMomentAs(eventStartDate);
+      case RepeatFrequency.daily:
+        return _isDailyRecurrence(
+          currentDate: currentDate,
+          recurrenceSettings: recurrenceSettings,
+        );
+      case RepeatFrequency.weekly:
+        return _isWeeklyRecurrence(
+          currentDate: currentDate,
+          recurrenceSettings: recurrenceSettings,
+        );
+      case RepeatFrequency.monthly:
+        return _isMonthlyRecurrence(
+          currentDate: currentDate,
+          startDate: eventStartDate,
+          recurrenceSettings: recurrenceSettings,
+        );
+      case RepeatFrequency.yearly:
+        return _isYearlyRecurrence(
+            currentDate: currentDate,
+            startDate: eventStartDate,
+            recurrenceSettings: recurrenceSettings);
+    }
+  }
+
+  //#endregion
+
   //#region Data Fetch Methods
   List<CalendarEventData<T>> getEventsOnDay(DateTime date,
       {bool includeFullDayEvents = true}) {
@@ -320,6 +548,7 @@ class CalendarData<T extends Object?> {
       events.addAll(_singleDayEvents[date]!);
     }
 
+    // TODO(Shubham): Add recurrence support for ranging events
     for (final rangingEvent in _rangingEventList) {
       if (rangingEvent.occursOnDate(date)) {
         events.add(rangingEvent);
@@ -330,6 +559,42 @@ class CalendarData<T extends Object?> {
       events.addAll(getFullDayEvent(date));
     }
 
+    // Add single day recurring events
+    final recurringEvents = getRecurringEventsOnDay(date)
+        .where((event) => (!event.isFullDayEvent && !event.isRangingEvent))
+        .toList();
+    events.addAll(recurringEvents);
+    return events;
+  }
+
+  /// Returns repeated events on given date.
+  List<CalendarEventData<T>> getRecurringEventsOnDay(DateTime date) {
+    final events = <CalendarEventData<T>>[];
+
+    //  Iterate through all repeated events and skips
+    //  if the given date is before start date of repeating event
+    //  or if the date is in excluded list of dates.
+    //  We do not need to handle Recurrence for it.
+    for (final event in _recurringEventsList) {
+      // recurrenceSettings is force casted because events in
+      // _recurringEventsList are added only if recurrenceSettings exists
+      final recurrenceSettings = event.recurrenceSettings!;
+
+      if (date.isBefore(event.date) || _isExcluded(recurrenceSettings, date)) {
+        continue;
+      }
+
+      final isRecurrence = _handleRecurrence(
+        currentDate: date,
+        eventStartDate: event.date,
+        eventEndDate: event.endDate,
+        recurrenceSettings: recurrenceSettings,
+      );
+
+      if (isRecurrence) {
+        events.add(event);
+      }
+    }
     return events;
   }
 
@@ -342,6 +607,12 @@ class CalendarData<T extends Object?> {
         events.add(event);
       }
     }
+
+    // Add full day recurring event
+    final recurringEvents = getRecurringEventsOnDay(date)
+        .where((event) => event.isFullDayEvent)
+        .toList();
+    events.addAll(recurringEvents);
     return events;
   }
   //#endregion